backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx==0.25.0
redis==5.0.0
prometheus-client==0.18.0
python-multipart==0.0.6
PyJWT==2.12.1
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiosqlite==0.19.0
email-validator==2.1.0
pydantic[email]==2.4.2
bcrypt==4.0.1
